+++ src/compas_view2/values/boolvalue.py
@@ -1,0 +1,7 @@
+from .value import Value
+
+
+class BoolValue(Value):
+
+    def __init__(self, value: bool):
+        super().__init__(value, bool)